Moved IComponentDebug component changes into _RMC folder
