+++ Content.Shared/Actions/Components/ActionsComponent.cs
@@ -1,4 +1,3 @@
-using System.Linq;
 using Content.Shared.Actions;
 using Robust.Shared.GameStates;
 using Robust.Shared.Serialization;
@@ -9,7 +8,7 @@ namespace Content.Shared.Actions.Components;
 /// Lets the player controlling this entity use actions.
 /// </summary>
 [RegisterComponent, NetworkedComponent, Access(typeof(SharedActionsSystem))]
-public sealed partial class ActionsComponent : Component, IComponentDebug
+public sealed partial class ActionsComponent : Component
 {
     /// <summary>
     /// List of actions currently granted to this entity.
@@ -17,13 +16,6 @@ public sealed partial class ActionsComponent : Component, IComponentDebug
     /// </summary>
     [DataField]
     public HashSet<EntityUid> Actions = new();
-
-    public string GetDebugString()
-    {
-        return $"""
-            Actions Count: {Actions.Count}
-            """;
-    }
 }
 
 [Serializable, NetSerializable]
