+++ Content.Shared/Atmos/Components/FlammableComponent.cs
@@ -1,4 +1,3 @@
-using System.Diagnostics.Metrics;
 using Content.Shared.Alert;
 using Content.Shared.Damage;
 using Robust.Shared.GameStates;
@@ -8,7 +7,7 @@ using Robust.Shared.Prototypes;
 namespace Content.Shared.Atmos.Components
 {
     [RegisterComponent, NetworkedComponent, AutoGenerateComponentState]
-    public sealed partial class FlammableComponent : Component, IComponentDebug
+    public sealed partial class FlammableComponent : Component
     {
         [DataField, AutoNetworkedField]
         public bool Resisting;
@@ -99,17 +98,5 @@ namespace Content.Shared.Atmos.Components
 
         [DataField, AutoNetworkedField]
         public DamageSpecifier? TileDamage;
-
-        public string GetDebugString()
-        {
-            return $"""
-            FireSpread: {FireSpread}
-            CanResistFire: {CanResistFire}
-            FireSpread: {FireSpread}
-            Damage: {Damage}
-            ResistStacks: {ResistStacks}
-            ResistDuration: {ResistDuration.TotalSeconds}
-            """;
-        }
     }
 }
