+++ Content.Shared/_RMC14/Atmos/Components/FlammableComponent.cs
@@ -1,0 +1,17 @@
+// ReSharper disable CheckNamespace
+namespace Content.Shared.Atmos.Components;
+
+public sealed partial class FlammableComponent : IComponentDebug
+{
+    public string GetDebugString()
+    {
+        return $"""
+            FireSpread: {FireSpread}
+            CanResistFire: {CanResistFire}
+            FireSpread: {FireSpread}
+            Damage: {Damage}
+            ResistStacks: {ResistStacks}
+            ResistDuration: {ResistDuration.TotalSeconds}
+            """;
+    }
+}
